refactor: small SCSS and code style refactoring
